+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Note: For changes to the API, see https://shopify.dev/changelog?filter=api
 
 ## Unreleased
 
+- [#1183](https://github.com/Shopify/shopify-api-ruby/pull/1189) Added string array support for fields parameter in Webhook::Registry
 - [#1192](https://github.com/Shopify/shopify-api-ruby/pull/1192) Fix no results when fetching CustomerAddress.all
 
 ## 13.1.0